test: move user-defined transform test to transform_multi
